token

+++ Life-Ecommerce/Controllers/ProductController.cs
@@ -1,6 +1,7 @@
 using Application.BackgroundJobs.ProductAnalytics;
 using Application.Services.ImageStorage;
 using Application.Services.Product;
+using Application.Services.TokenService;
 using Domain.DTOs.Product;
 using Microsoft.AspNetCore.Mvc;
 using Presistence.Repositories.ProductAnalytics;
@@ -15,13 +16,15 @@ public class ProductController : ControllerBase
     private readonly IStorageService _storageService;
     private readonly ILogger<ProductController> _logger;
     private readonly IProductAnalyticsService _productAnalyticsService;
+    private readonly TokenHelper _tokenHelper;
 
-    public ProductController(IProductService productService, IStorageService storageService, ILogger<ProductController> logger, IProductAnalyticsService productAnalyticsService)
+    public ProductController(IProductService productService, IStorageService storageService, ILogger<ProductController> logger, IProductAnalyticsService productAnalyticsService, TokenHelper token)
     {
         _productService = productService;
         _logger = logger;
         _storageService = storageService;
         _productAnalyticsService = productAnalyticsService;
+        _tokenHelper = token;
 
     }
 
@@ -68,6 +71,15 @@ public class ProductController : ControllerBase
     [Consumes("multipart/form-data")]
     public async Task<ActionResult> AddProduct([FromForm] CreateProductDto createProductDto)
     {
+        var userRole = _tokenHelper.GetUserRole();
+        if (userRole == null)
+        {
+            return Unauthorized("You are not logged in.");
+        }
+        else if (userRole != "Admin" && userRole != "SuperAdmin")
+        {
+            return Unauthorized("You are not authorized to perform this action.");
+        }
         await _productService.AddProductAsync(createProductDto);
 
         return Ok("Product added succesfully to the database and indexed to Elastic");
@@ -76,6 +88,15 @@ public class ProductController : ControllerBase
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateProduct(int id, [FromForm] UpdateProductDto updateProductDto)
     {
+        var userRole = _tokenHelper.GetUserRole();
+        if (userRole == null)
+        {
+            return Unauthorized("You are not logged in.");
+        }
+        else if (userRole != "Admin" && userRole != "SuperAdmin")
+        {
+            return Unauthorized("You are not authorized to perform this action.");
+        }
 
         var response = await _productService.UpdateProductAsync(id, updateProductDto);
         return Ok(response);
@@ -84,6 +105,15 @@ public class ProductController : ControllerBase
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteProduct(int id)
     {
+        var userRole = _tokenHelper.GetUserRole();
+        if (userRole == null)
+        {
+            return Unauthorized("You are not logged in.");
+        }
+        else if (userRole != "Admin" && userRole != "SuperAdmin")
+        {
+            return Unauthorized("You are not authorized to perform this action.");
+        }
         var result = await _productService.DeleteProductAsync(id);
 
         if (!result)
@@ -133,6 +163,15 @@ public class ProductController : ControllerBase
     [HttpDelete("softdelete/{productId}")]
     public async Task<ActionResult> SoftDeleteProduct(int productId)
     {
+        var userRole = _tokenHelper.GetUserRole();
+        if (userRole == null)
+        {
+            return Unauthorized("You are not logged in.");
+        }
+        else if (userRole != "Admin" && userRole != "SuperAdmin")
+        {
+            return Unauthorized("You are not authorized to perform this action.");
+        }
         await _productService.SoftDeleteProduct(productId);
         return NoContent();
     }
@@ -141,6 +180,15 @@ public class ProductController : ControllerBase
     [HttpPost("discount/{productId}")]
     public async Task<ActionResult> AddDiscountToProduct(int productId, decimal discount, DateTime ExpiryDate)
     {
+        var userRole = _tokenHelper.GetUserRole();
+        if (userRole == null)
+        {
+            return Unauthorized("You are not logged in.");
+        }
+        else if (userRole != "Admin" && userRole != "SuperAdmin")
+        {
+            return Unauthorized("You are not authorized to perform this action.");
+        }
         await _productService.AddDiscountToProduct(productId, discount, ExpiryDate);
         return Ok("Discount added successfully.");
     }
